chore: split the mock quiz tag and tag schema model file in the tags/models/
